requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
python-multipart
passlib[bcrypt]
python-jose[cryptography]
Pillow
numpy
